issue #35 (DASHUP-107): implemented test for change user name
